(*) restructure project
the top-level CMakeLists now is just a wrapper for IDEs (CLion...)

the library code can now be found in subdirectory aide and can
be built standalone. If the library is build standalone,
it has not enabled any project warnings, sanitizers, etc.

When the library is prepared to be installed also the tests
can be built standalone, which enables to test against a
installed target to also check if the installation works

these steps follow the tipps from Mateusz Pusz:
https://www.youtube.com/watch?v=S4QSKLXdTtA