(*) do not explicitly build shared libraries
there is a bug on windows, where the logger test segfaults
https://github.com/mrpilot2/aide/issues/6

therefore now we can decide from outside if we would like to build
static or shared libraries

Also there is an issue with CMake when we link interface libraries
to a static library and want to export the static libraries.
Then CMake want to also export the interface libraries
which I don't want to export because it's just sanitizer and
coverage settings.
The workaround here is to add the libraries as INTERFACE IMPORTED
see https://gitlab.kitware.com/cmake/cmake/issues/15415#note_634114
and https://gitlab.kitware.com/cmake/cmake/issues/17357